remake db maker

+++ database/migrations/2022_05_19_142037_create_balitas_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('jk');
             $table->integer('usia');
-            $table->integer('bb');
+            $table->float('bb');
             $table->timestamps();
         });
     }
